refactor(privacy): extract WriteOnlyStreamDecorator for export streams
Pull the shared write-only Stream plumbing out of CountingStream,
Sha256ComputingStream, and WriteCountingStream into a single
WriteOnlyStreamDecorator base (~242 lines removed); each stream now overrides
only its write-side behaviour.
